+++ packages/xml-views-validation/test/unit/validators/attributes/unknown-attribute-key.test.ts
@@ -5,6 +5,8 @@ import { UI5SemanticModel } from "@ui5-language-assistant/semantic-model-types";
 import {
   generateModel,
   expectExists,
+  DEFAULT_UI5_VERSION,
+  getFallbackPatchVersions,
 } from "@ui5-language-assistant/test-utils";
 import { generate } from "@ui5-language-assistant/semantic-model";
 import { validators } from "../../../../src/api";
@@ -21,7 +23,9 @@ describe("the unknown attribute name validation", () => {
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.61",
+      version: (
+        await getFallbackPatchVersions()
+      ).SAPUI5 as typeof DEFAULT_UI5_VERSION,
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);