+++ packages/language-server/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 4.0.42
+
+### Patch Changes
+
+- 0186685: Fallback to the latest supported patch version
+- Updated dependencies [0186685]
+  - @ui5-language-assistant/xml-views-completion@4.0.23
+  - @ui5-language-assistant/xml-views-validation@4.0.18
+  - @ui5-language-assistant/xml-views-tooltip@4.0.23
+  - @ui5-language-assistant/semantic-model@4.0.16
+  - @ui5-language-assistant/logic-utils@4.0.17
+  - @ui5-language-assistant/context@4.0.23
+  - @ui5-language-assistant/binding@1.0.25
+  - @ui5-language-assistant/fe@4.0.24
+  - @ui5-language-assistant/xml-views-quick-fix@4.0.17
+
 ## 4.0.41
 
 ### Patch Changes