+++ packages/xml-views-validation/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 4.0.18
+
+### Patch Changes
+
+- 0186685: Fallback to the latest supported patch version
+- Updated dependencies [0186685]
+  - @ui5-language-assistant/logic-utils@4.0.17
+
 ## 4.0.17
 
 ### Patch Changes